#35: restructure project like #43
