+++ lib/restore-utils/data-dictionary.js
@@ -11,7 +11,7 @@ const {
   findDataDictionaryReport,
   deleteExistingDDOrWebAPIReport,
   sleep
-} = require('../../data-access/cert-api-client');
+} = require('../misc/data-access/cert-api-client');
 
 const { processLookupResourceMetadataFiles } = require('reso-certification-etl');
 const { checkFileExists } = require('../../common');
