+++ lib/restore-utils/web-api-core.js
@@ -10,7 +10,7 @@ const {
   getSystemsMap,
   deleteExistingDDOrWebAPIReport,
   sleep
-} = require('../../data-access/cert-api-client');
+} = require('../misc/data-access/cert-api-client');
 
 const OVERWRITE_DELAY_S = 10;
 
@@ -70,6 +70,7 @@ const isValidUrl = url => {
  * @param {string} options.url Cert API base URL.
  * @param {string} options.recipients Comma seperated string of recipient uoi's.
  * @param {string} options.system System name for the provider.
+ * @param {boolean} options.deleteExistingReport Flag to force delete an existing web API report
  * @throws Error if path is not a valid S3 or local path
  */
 const syncWebApi = async (options = {}) => {
